chore: refactoring entity extraction on ws-extractor
